moved old add screens to dev folder
Took 4 minutes
